Only store configuration hashes instead of full configuration in managed mode

See merge request chromaway/postchain!1194

+++ postchain-server/src/main/kotlin/net/postchain/server/service/PostchainService.kt
@@ -12,6 +12,7 @@ import net.postchain.base.withReadConnection
 import net.postchain.base.withWriteConnection
 import net.postchain.common.BlockchainRid
 import net.postchain.common.exception.NotFound
+import net.postchain.common.exception.UserMistake
 import net.postchain.core.BadDataException
 import net.postchain.crypto.KeyPair
 import net.postchain.crypto.PrivKey
@@ -130,6 +131,11 @@ class PostchainService(private val nodeProvider: NodeProvider) {
         stopBlockchain(chainId)
 
         withWriteConnection(postchainNode.postchainContext.sharedStorage, chainId) { ctx ->
+            val dependentChains = BlockchainApi.getDependentChains(ctx)
+            if (dependentChains.isNotEmpty()) {
+                throw UserMistake("Blockchain may not be deleted due to the following dependent chains: ${dependentChains.joinToString(", ")}")
+            }
+
             BlockchainApi.deleteBlockchain(ctx)
             true
         }
